Add files via upload
Here was our first assignment, looking into the intro to Bayesian Theory. The TLDR is that in the world of statistics we need to count, and those counts make probabilities. When we point our telescope at the possible outcomes, we can work backwards to make a hypothesis (and test it) on what outcomes are more likely. All we need to answer our hypothesis is a bit of test data.